+++ CHANGELOG.md
@@ -61,7 +61,7 @@ Malli is in [alpha](README.md#alpha).
 
 * new `malli.instrument` and `malli.dev` for instrumenting function Vars (e.g. `defn`s), see [the guide](docs/function-schemas.md).
 * new `malli.plantuml` namespace for [PlantUML generation](README.md#plantuml)
-* new `malli.generator/check` and `malli.generator/check!` for generative testing of functions and `defn`s.
+* new `malli.generator/check` for generative testing of functions and `defn`s.
 * new `malli.core/parent`
 * `:map-of` supports `:min` and `:max` properties
 * Collection Schemas emit correct JSON Schema min & max declarations